Implement the new encoding with 64bit size and expire time in milliseconds (#1342)
See proposal in #1033, after this change, these old data can still be readable/writable, but all new data will be written via the new encoding if `ENABLE_NEW_ENCODING` is turned on (currently default OFF).

NOTE: Although the encoding now support 64bit size, maybe some place in code still use int32_t/int so it cannot work well in large number of items more than 32bit. We can fix them after this PR.

+++ src/storage/redis_db.h
@@ -35,10 +35,10 @@ class Database {
   rocksdb::Status GetMetadata(RedisType type, const Slice &ns_key, Metadata *metadata);
   rocksdb::Status GetRawMetadata(const Slice &ns_key, std::string *bytes);
   rocksdb::Status GetRawMetadataByUserKey(const Slice &user_key, std::string *bytes);
-  rocksdb::Status Expire(const Slice &user_key, int timestamp);
+  rocksdb::Status Expire(const Slice &user_key, uint64_t timestamp);
   rocksdb::Status Del(const Slice &user_key);
   rocksdb::Status Exists(const std::vector<Slice> &keys, int *ret);
-  rocksdb::Status TTL(const Slice &user_key, int *ttl);
+  rocksdb::Status TTL(const Slice &user_key, int64_t *ttl);
   rocksdb::Status Type(const Slice &user_key, RedisType *type);
   rocksdb::Status Dump(const Slice &user_key, std::vector<std::string> *infos);
   rocksdb::Status FlushDB();
